General Cleanup
- moving contexts closer to where they belong
- split GameContext into
  - GameContext
  - gameReducer
  - types
- moved some components from /src/screens/components to /src/components
- deleted /src/screens/components
