--- src/screens/game/util/useFrameStep.tsx
@@ -1,3 +1,0 @@
-// todo
-
-export function useFrameStep() {}
